+++ DependencyInversionPrinciple/Database.java
@@ -1,10 +1,11 @@
 package DependencyInversionPrinciple;
 
+import DependencyInversionPrinciple.model.Employee;
+
 public class Database implements DatabaseAccess{
 
     @Override
     public void save(Employee employee) {
         // Code to save employee to the database
     }
-
 }
