+++ DependencyInversionPrinciple/DatabaseAccess.java
@@ -1,5 +1,7 @@
 package DependencyInversionPrinciple;
 
+import DependencyInversionPrinciple.model.Employee;
+
 public interface DatabaseAccess {
     public void save(Employee employee);   
 }
